requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
Django==3.2.5
django-filter==2.4.0
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-sendgrid-v5==0.9.1
future==0.18.2
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
psycopg2==2.9.1
PyJWT==1.7.1
python-http-client==3.3.2
pytz==2021.1
Rx==1.6.1
sendgrid==6.7.1
singledispatch==3.6.2
six==1.16.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
text-unidecode==1.3
